requirements.txt
langchain
langchain_core
llama_parse
pandas
pdfplumber
PyPDF2
python-dotenv
python_Levenshtein
Requests
tabula_py
tqdm
marker-pdf
tiktoken
langchain-ollama
unstructured==0.12.5